Rename: auth 관련 유틸 함수 파일 정리
