+++ src/pages/SignUp.tsx
@@ -1,7 +1,7 @@
 import { ChangeEvent, useState } from 'react';
 import { useNavigate } from 'react-router';
 import { UserSignInput } from '../types/type';
-import { isValidEmail, isValidPassword } from '../utils/utils';
+import { isValidEmail, isValidPassword } from '../utils/auth';
 import { useAuthContext } from '../hooks/useAuthContext';
 import { styled } from 'styled-components';
 import Button from '../components/Button';
